Ghost
Ghost class with seekX, seekY, and move methods. Creatures have to have
a reference the activity they’re in so made that part of the
constructor. Also added a getMainCharacter() method to the activity
GamePractice. Implemented methods in MainCharacter so that it could
compile but didn’t flesh them out.